Name every spec file after the prod file it covers
Spec file names had drifted into two families: some named after their prod
module (`helpers.spec.ts`, `bot-strategy.spec.ts`), others after a concern
(`legality.spec.ts`, `end-of-game.spec.ts`, `optimality.spec.ts`,
`bot-behaviour.spec.ts`). The concern-named ones all tested exports of their
folder's main game file, so the second family carried no information the first
didn't — it just made "where do I add a test for X" ambiguous.

Every spec is now `<prod-file>.spec.ts` next to a `<prod-file>` it imports.
Concern-named specs were renamed where the target was free, and merged into
the existing game spec otherwise (validator describes first, then end of game).

Two splits, both to follow the folder's own structure:

- `chocolate-breaking.spec.ts` exclusively tested `helpers.ts` and
  `bot-strategy.ts`, so it became `helpers.spec.ts` + `bot-strategy.spec.ts`,
  and the old `end-of-game.spec.ts` (which does test the game file's `moves`)
  took the `chocolate-breaking.spec.ts` name.
- `policeman-thief-ab/legality.spec.ts` covered both the graph predicates in
  `helpers.ts` and the moves' validators, so the graph half moved to a new
  `helpers.spec.ts`, matching the sibling `policeman-thief-c` folder.

192 spec files down to 173, with the same 1197 tests passing. Merges reconciled
colliding local `board` helpers by reusing the target file's existing builder
where the two were equivalent. No prod code changed beyond two comments naming
the renamed `gameList.spec.ts`.
