+++ src/components/games/discs-flip-or-remove/discs-flip-or-remove.spec.ts
@@ -2,15 +2,53 @@ import { describe, it, expect } from 'vitest';
 import { mapValues } from 'lodash';
 import { type GameMoves } from '../../strategy-game-factory';
 import { makeCtx } from '../../../test-utils';
-import { moves as gameMoves, smartBotStrategy } from './discs-flip-or-remove';
+import { isFlipAllowed, isRemovalAllowed, moves as gameMoves, smartBotStrategy } from './discs-flip-or-remove';
+
+// board[0] = blue discs, board[1] = red discs.
+type Board = [number, number];
+
+describe('isRemovalAllowed', () => {
+  it('accepts taking one or two blue discs', () => {
+    expect(isRemovalAllowed([3, 4], 1)).toBe(true);
+    expect(isRemovalAllowed([3, 4], 2)).toBe(true);
+  });
+
+  it('refuses any other count', () => {
+    expect(isRemovalAllowed([3, 4], 0)).toBe(false);
+    expect(isRemovalAllowed([3, 4], 3)).toBe(false);
+    expect(isRemovalAllowed([3, 4], -1)).toBe(false);
+  });
+
+  it('refuses taking more blue discs than there are', () => {
+    expect(isRemovalAllowed([1, 4], 2)).toBe(false);
+    expect(isRemovalAllowed([1, 4], 1)).toBe(true);
+    expect(isRemovalAllowed([0, 4], 1)).toBe(false);
+  });
+});
+
+describe('isFlipAllowed', () => {
+  it('accepts flipping one or two red discs', () => {
+    expect(isFlipAllowed([3, 4], 1)).toBe(true);
+    expect(isFlipAllowed([3, 4], 2)).toBe(true);
+  });
+
+  it('refuses flipping more red discs than there are', () => {
+    expect(isFlipAllowed([3, 1], 2)).toBe(false);
+    expect(isFlipAllowed([3, 1], 1)).toBe(true);
+    expect(isFlipAllowed([3, 0], 1)).toBe(false);
+  });
+
+  it('counts each pile separately — a full blue pile does not license a red flip', () => {
+    expect(isFlipAllowed([9, 0], 1)).toBe(false);
+    expect(isRemovalAllowed([0, 9], 1)).toBe(false);
+  });
+});
 
 // Exhaustive optimality check for both variants (max 6 and max 10 discs).
-// Board = [blue, red]. Moves: remove 1-2 blue, or flip 1-2 red into blue.
+// Moves: remove 1-2 blue, or flip 1-2 red into blue.
 // A player who cannot move (board [0,0]) loses. We verify against an
 // independent minimax over the whole reachable state space that the bot,
 // whenever the mover can win, always moves to a losing-for-opponent position.
-
-type Board = [number, number];
 
 const key = ([b, r]: Board) => `${b},${r}`;
