+++ src/components/games/gameList.ts
@@ -9,7 +9,7 @@ The two files are kept apart on purpose. Merging them would drag the whole
 game-component graph into every metadata consumer's bundle. This file is imported
 ~a dozen places; `index.ts` is imported in exactly one runtime place — the router.
 `app.tsx` pairs them: it loops these keys and looks each up in the `index.ts`
-namespace. `game-list-integrity.spec.ts` guards that the two stay 1-to-1 in sync.
+namespace. `gameList.spec.ts` guards that the two stay 1-to-1 in sync.
 */
 import type { I18nString } from '../../language';
