+++ src/components/games/number-covering/number-covering.spec.ts
@@ -1,11 +1,56 @@
-import { moves } from './number-covering';
+import { isCoveringAllowed, getOptimalSmartBotMove, moves, COVERED } from './number-covering';
 import { makeCtx } from '../../../test-utils';
+
+const meta = { ctx: makeCtx() };
+
+describe('isCoveringAllowed', () => {
+  const board = [1, 2, COVERED, 4, 5];
+
+  it('allows covering a number still showing', () => {
+    expect([1, 2, 4, 5].every(number => isCoveringAllowed(board, number))).toBe(true);
+  });
+
+  it('rejects a number already covered', () => {
+    expect(isCoveringAllowed(board, 3)).toBe(false);
+  });
+
+  it('rejects a number not on the table', () => {
+    expect(isCoveringAllowed(board, 6)).toBe(false);
+    expect(isCoveringAllowed(board, 0)).toBe(false);
+    expect(isCoveringAllowed(board, 2.5)).toBe(false);
+  });
+});
+
+describe('smart bot', () => {
+  // 1..8 with 7 covered -> 4 evens (2,4,6,8) and 3 odds (1,3,5) remain.
+  const fourEvensThreeOdds = [1, 2, 3, 4, 5, 6, COVERED, 8];
+  // 1..8 with 6 and 8 covered -> 2 evens (2,4) and 4 odds (1,3,5,7) remain.
+  const twoEvensFourOdds = [1, 2, 3, 4, 5, COVERED, 7, COVERED];
+
+  it('as first player covers from the smaller parity class', () => {
+    // odds are the minority -> covers an odd
+    expect([1, 3, 5]).toContain(getOptimalSmartBotMove({ board: fourEvensThreeOdds, currentPlayer: 0 }));
+    // evens are the minority -> covers an even
+    expect([2, 4]).toContain(getOptimalSmartBotMove({ board: twoEvensFourOdds, currentPlayer: 0 }));
+  });
+
+  it('as second player covers from the larger parity class', () => {
+    // evens are the majority -> covers an even
+    expect([2, 4, 6, 8]).toContain(getOptimalSmartBotMove({ board: fourEvensThreeOdds, currentPlayer: 1 }));
+    // odds are the majority -> covers an odd
+    expect([1, 3, 5, 7]).toContain(getOptimalSmartBotMove({ board: twoEvensFourOdds, currentPlayer: 1 }));
+  });
+
+  it('plays a legal move when the parities are balanced', () => {
+    const numbers = [1, 2, 3, 4, 5, 6, 7, 8]; // 4 evens, 4 odds
+    expect(numbers).toContain(getOptimalSmartBotMove({ board: numbers, currentPlayer: 0 }));
+    expect(numbers).toContain(getOptimalSmartBotMove({ board: numbers, currentPlayer: 1 }));
+  });
+});
 
 // The game runs until two numbers are left, and the *parity of their sum*
 // decides it: even hands it to the first player, odd to the second. Nothing
 // about whose turn it was matters, which is what these tests pin.
-const meta = { ctx: makeCtx() };
-
 describe('end of game', () => {
   it('gives an even remaining sum to the first player', () => {
     // covering 2 leaves 1 and 3
