+++ src/components/games/six-fields-circle/six-fields-circle.spec.ts
