+++ src/components/games/ten-digit-number/ten-digit-number.spec.ts
@@ -1,8 +1,6 @@
-import { moves } from './ten-digit-number';
+import { isDigitChoiceAllowed, moves } from './ten-digit-number';
 import { makeCtx } from '../../../test-utils';
 
-// The two players build one ten-digit number together; Bob (player 1) wins iff
-// it is divisible by 9. Neither the mover nor the turn order enters into it.
 const meta = { ctx: makeCtx() };
 
 const board = (digits: number[]) => ({
@@ -10,6 +8,31 @@ const board = (digits: number[]) => ({
   sumMod9: digits.reduce((acc, d) => acc + d, 0) % 9
 });
 
+describe('isDigitChoiceAllowed', () => {
+  const afterDigits = (count: number) => board(Array(count).fill(1));
+
+  it('accepts each of the six offered digits', () => {
+    for (const d of [1, 2, 3, 4, 5, 6]) expect(isDigitChoiceAllowed(afterDigits(0), d)).toBe(true);
+  });
+
+  it('refuses a digit outside the offered six', () => {
+    expect(isDigitChoiceAllowed(afterDigits(0), 0)).toBe(false);
+    expect(isDigitChoiceAllowed(afterDigits(0), 7)).toBe(false);
+    expect(isDigitChoiceAllowed(afterDigits(0), 9)).toBe(false);
+    expect(isDigitChoiceAllowed(afterDigits(0), -1)).toBe(false);
+  });
+
+  it('accepts a digit up to the last free slot', () => {
+    expect(isDigitChoiceAllowed(afterDigits(9), 3)).toBe(true);
+  });
+
+  it('refuses any digit once all ten are written', () => {
+    expect(isDigitChoiceAllowed(afterDigits(10), 3)).toBe(false);
+  });
+});
+
+// The two players build one ten-digit number together; Bob (player 1) wins iff
+// it is divisible by 9. Neither the mover nor the turn order enters into it.
 describe('end of game', () => {
   it('gives the tenth digit to Bob when the sum lands on a multiple of 9', () => {
     // nine 1s sum to 9; adding a 9 keeps the total ≡ 0 (mod 9)
